separate global actions (#194802)
