+++ src/vs/workbench/browser/parts/activitybar/activitybarPart.ts
@@ -5,48 +5,38 @@
 
 import 'vs/css!./media/activitybarpart';
 import { localize } from 'vs/nls';
-import { ActionsOrientation, ActionBar } from 'vs/base/browser/ui/actionbar/actionbar';
-import { GLOBAL_ACTIVITY_ID, IActivity, ACCOUNTS_ACTIVITY_ID } from 'vs/workbench/common/activity';
+import { ActionsOrientation } from 'vs/base/browser/ui/actionbar/actionbar';
 import { Part } from 'vs/workbench/browser/part';
-import { GlobalActivityActionViewItem, AccountsActivityActionViewItem } from 'vs/workbench/browser/parts/activitybar/activitybarActions';
-import { IBadge, NumberBadge } from 'vs/workbench/services/activity/common/activity';
+import { IBadge } from 'vs/workbench/services/activity/common/activity';
 import { IWorkbenchLayoutService, Parts, Position } from 'vs/workbench/services/layout/browser/layoutService';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { IDisposable, toDisposable, DisposableStore } from 'vs/base/common/lifecycle';
+import { IDisposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { ToggleActivityBarVisibilityAction, ToggleSidebarPositionAction } from 'vs/workbench/browser/actions/layoutActions';
 import { IThemeService, IColorTheme } from 'vs/platform/theme/common/themeService';
 import { ACTIVITY_BAR_BACKGROUND, ACTIVITY_BAR_BORDER, ACTIVITY_BAR_FOREGROUND, ACTIVITY_BAR_ACTIVE_BORDER, ACTIVITY_BAR_BADGE_BACKGROUND, ACTIVITY_BAR_BADGE_FOREGROUND, ACTIVITY_BAR_INACTIVE_FOREGROUND, ACTIVITY_BAR_ACTIVE_BACKGROUND, ACTIVITY_BAR_DRAG_AND_DROP_BORDER } from 'vs/workbench/common/theme';
 import { contrastBorder } from 'vs/platform/theme/common/colorRegistry';
 import { addDisposableListener, EventType, isAncestor } from 'vs/base/browser/dom';
-import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
-import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
-import { ICompositeBarColors, ActivityAction, ICompositeActivity, IActivityHoverOptions } from 'vs/workbench/browser/parts/compositeBarActions';
+import { ICompositeBarColors, IActivityHoverOptions } from 'vs/workbench/browser/parts/compositeBarActions';
 import { assertIsDefined } from 'vs/base/common/types';
 import { CustomMenubarControl } from 'vs/workbench/browser/parts/titlebar/menubarControl';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { getMenuBarVisibility } from 'vs/platform/window/common/window';
-import { Codicon } from 'vs/base/common/codicons';
-import { ThemeIcon } from 'vs/base/common/themables';
 import { Separator, toAction } from 'vs/base/common/actions';
 import { StandardKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { KeyCode } from 'vs/base/common/keyCodes';
-import { registerIcon } from 'vs/platform/theme/common/iconRegistry';
 import { HoverPosition } from 'vs/base/browser/ui/hover/hoverWidget';
 import { GestureEvent } from 'vs/base/browser/touch';
 import { IPaneCompositePart } from 'vs/workbench/browser/parts/paneCompositePart';
-import { IUserDataProfileService } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
-import { DEFAULT_ICON } from 'vs/workbench/services/userDataProfile/common/userDataProfileIcons';
 import { PaneCompositeBar } from 'vs/workbench/browser/parts/paneCompositeBar';
+import { GlobalCompositeBar } from 'vs/workbench/browser/parts/globalCompositeBar';
+import { IStorageService } from 'vs/platform/storage/common/storage';
 
 export class ActivitybarPart extends Part {
 
 	private static readonly ACTION_HEIGHT = 48;
-	private static readonly ACCOUNTS_ACTION_INDEX = 0;
 
 	static readonly pinnedViewContainersKey = 'workbench.activity.pinnedViewlets2';
 	static readonly placeholderViewContainersKey = 'workbench.activity.placeholderViewlets';
-
-	private static readonly ACCOUNTS_ICON = registerIcon('accounts-view-bar-icon', Codicon.account, localize('accountsViewBarIcon', "Accounts icon in the view bar."));
 
 	//#region IView
 
@@ -64,14 +54,7 @@ export class ActivitybarPart extends Part {
 
 	private compositeBarContainer: HTMLElement | undefined;
 	private readonly compositeBar: PaneCompositeBar;
-
-	private globalActivityAction: ActivityAction | undefined;
-	private globalActivityActionBar: ActionBar | undefined;
-	private globalActivitiesContainer: HTMLElement | undefined;
-	private readonly globalActivity: ICompositeActivity[] = [];
-
-	private accountsActivityAction: ActivityAction | undefined;
-	private readonly accountsActivity: ICompositeActivity[] = [];
+	private readonly globalCompositeBar: GlobalCompositeBar;
 
 	private readonly keyboardNavigationDisposables = this._register(new DisposableStore());
 
@@ -80,13 +63,12 @@ export class ActivitybarPart extends Part {
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 		@IWorkbenchLayoutService layoutService: IWorkbenchLayoutService,
 		@IThemeService themeService: IThemeService,
-		@IStorageService private readonly storageService: IStorageService,
-		@IExtensionService private readonly extensionService: IExtensionService,
+		@IStorageService storageService: IStorageService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
-		@IUserDataProfileService private readonly userDataProfileService: IUserDataProfileService,
 	) {
 		super(Parts.ACTIVITYBAR_PART, { hasTitle: false }, themeService, storageService, layoutService);
 		this.compositeBar = this.createCompositeBar();
+		this.globalCompositeBar = this._register(instantiationService.createInstance(GlobalCompositeBar, () => this.compositeBar.getContextMenuActions(), (theme: IColorTheme) => this.getActivitybarItemColors(theme), this.getActivityHoverOptions()));
 		this.registerListeners();
 	}
 
@@ -113,9 +95,9 @@ export class ActivitybarPart extends Part {
 					}
 				}
 
-				// Accounts
+				// Global Composite Bar
 				actions.push(new Separator());
-				actions.push(toAction({ id: 'toggleAccountsVisibility', label: localize('accounts', "Accounts"), checked: this.accountsVisibilityPreference, run: () => this.accountsVisibilityPreference = !this.accountsVisibilityPreference }));
+				actions.push(...this.globalCompositeBar.getContextMenuActions());
 				actions.push(new Separator());
 
 				// Toggle Sidebar
@@ -137,13 +119,6 @@ export class ActivitybarPart extends Part {
 	}
 
 	private registerListeners(): void {
-
-		// Extension registration
-		const disposables = this._register(new DisposableStore());
-		this._register(this.extensionService.onDidRegisterExtensions(() => {
-			this.storageService.onDidChangeValue(StorageScope.PROFILE, AccountsActivityActionViewItem.ACCOUNTS_VISIBILITY_PREFERENCE_KEY, disposables)(() => this.toggleAccountsActivity(), this, disposables);
-		}));
-
 		// Register for configuration changes
 		this._register(this.configurationService.onDidChangeConfiguration(e => {
 			if (e.affectsConfiguration('window.menuBarVisibility')) {
@@ -157,83 +132,11 @@ export class ActivitybarPart extends Part {
 	}
 
 	showActivity(viewContainerOrActionId: string, badge: IBadge, clazz?: string, priority?: number): IDisposable {
-		if (viewContainerOrActionId === GLOBAL_ACTIVITY_ID) {
-			return this.showGlobalActivity(GLOBAL_ACTIVITY_ID, badge, clazz, priority);
+		const disposable = this.globalCompositeBar.showActivity(viewContainerOrActionId, badge, clazz, priority);
+		if (disposable) {
+			return disposable;
 		}
-
-		if (viewContainerOrActionId === ACCOUNTS_ACTIVITY_ID) {
-			return this.showGlobalActivity(ACCOUNTS_ACTIVITY_ID, badge, clazz, priority);
-		}
-
 		return this.compositeBar.showActivity(viewContainerOrActionId, badge, clazz, priority);
-	}
-
-	private showGlobalActivity(activityId: string, badge: IBadge, clazz?: string, priority?: number): IDisposable {
-		if (typeof priority !== 'number') {
-			priority = 0;
-		}
-
-		const activity: ICompositeActivity = { badge, clazz, priority };
-		const activityCache = activityId === GLOBAL_ACTIVITY_ID ? this.globalActivity : this.accountsActivity;
-
-		for (let i = 0; i <= activityCache.length; i++) {
-			if (i === activityCache.length) {
-				activityCache.push(activity);
-				break;
-			} else if (activityCache[i].priority <= priority) {
-				activityCache.splice(i, 0, activity);
-				break;
-			}
-		}
-		this.updateGlobalActivity(activityId);
-
-		return toDisposable(() => this.removeGlobalActivity(activityId, activity));
-	}
-
-	private removeGlobalActivity(activityId: string, activity: ICompositeActivity): void {
-		const activityCache = activityId === GLOBAL_ACTIVITY_ID ? this.globalActivity : this.accountsActivity;
-		const index = activityCache.indexOf(activity);
-		if (index !== -1) {
-			activityCache.splice(index, 1);
-			this.updateGlobalActivity(activityId);
-		}
-	}
-
-	private updateGlobalActivity(activityId: string): void {
-		const activityAction = activityId === GLOBAL_ACTIVITY_ID ? this.globalActivityAction : this.accountsActivityAction;
-		if (!activityAction) {
-			return;
-		}
-
-		const activityCache = activityId === GLOBAL_ACTIVITY_ID ? this.globalActivity : this.accountsActivity;
-		if (activityCache.length) {
-			const [{ badge, clazz, priority }] = activityCache;
-			if (badge instanceof NumberBadge && activityCache.length > 1) {
-				const cumulativeNumberBadge = this.getCumulativeNumberBadge(activityCache, priority);
-				activityAction.setBadge(cumulativeNumberBadge);
-			} else {
-				activityAction.setBadge(badge, clazz);
-			}
-		} else {
-			activityAction.setBadge(undefined);
-		}
-	}
-
-	private getCumulativeNumberBadge(activityCache: ICompositeActivity[], priority: number): NumberBadge {
-		const numberActivities = activityCache.filter(activity => activity.badge instanceof NumberBadge && activity.priority === priority);
-		const number = numberActivities.reduce((result, activity) => { return result + (<NumberBadge>activity.badge).number; }, 0);
-		const descriptorFn = (): string => {
-			return numberActivities.reduce((result, activity, index) => {
-				result = result + (<NumberBadge>activity.badge).getDescription();
-				if (index < numberActivities.length - 1) {
-					result = `${result}\n`;
-				}
-
-				return result;
-			}, '');
-		};
-
-		return new NumberBadge(number, descriptorFn);
 	}
 
 	private uninstallMenubar() {
@@ -283,10 +186,7 @@ export class ActivitybarPart extends Part {
 		this.compositeBarContainer = this.compositeBar.create(this.content);
 
 		// Global action bar
-		this.globalActivitiesContainer = document.createElement('div');
-		this.content.appendChild(this.globalActivitiesContainer);
-
-		this.createGlobalActivityActionBar(this.globalActivitiesContainer);
+		this.globalCompositeBar.create(this.content);
 
 		// Keyboard Navigation
 		this.registerKeyboardNavigationListeners();
@@ -312,7 +212,7 @@ export class ActivitybarPart extends Part {
 			this.keyboardNavigationDisposables.add(addDisposableListener(this.compositeBarContainer, EventType.KEY_DOWN, e => {
 				const kbEvent = new StandardKeyboardEvent(e);
 				if (kbEvent.equals(KeyCode.DownArrow) || kbEvent.equals(KeyCode.RightArrow)) {
-					this.globalActivityActionBar?.focus(true);
+					this.globalCompositeBar.focus();
 				} else if (kbEvent.equals(KeyCode.UpArrow) || kbEvent.equals(KeyCode.LeftArrow)) {
 					this.menuBar?.toggleFocus();
 				}
@@ -320,82 +220,12 @@ export class ActivitybarPart extends Part {
 		}
 
 		// Up arrow on global icons
-		if (this.globalActivitiesContainer) {
-			this.keyboardNavigationDisposables.add(addDisposableListener(this.globalActivitiesContainer, EventType.KEY_DOWN, e => {
-				const kbEvent = new StandardKeyboardEvent(e);
-				if (kbEvent.equals(KeyCode.UpArrow) || kbEvent.equals(KeyCode.LeftArrow)) {
-					this.compositeBar?.focus(this.getVisiblePaneCompositeIds().length - 1);
-				}
-			}));
-		}
-	}
-
-	private createGlobalActivityActionBar(container: HTMLElement): void {
-		this.globalActivityActionBar = this._register(new ActionBar(container, {
-			actionViewItemProvider: action => {
-				if (action.id === 'workbench.actions.manage') {
-					return this.instantiationService.createInstance(GlobalActivityActionViewItem, action as ActivityAction, () => this.compositeBar.getContextMenuActions(), (theme: IColorTheme) => this.getActivitybarItemColors(theme), this.getActivityHoverOptions());
-				}
-
-				if (action.id === 'workbench.actions.accounts') {
-					return this.instantiationService.createInstance(AccountsActivityActionViewItem, action as ActivityAction, () => this.compositeBar.getContextMenuActions(), (theme: IColorTheme) => this.getActivitybarItemColors(theme), this.getActivityHoverOptions());
-				}
-
-				throw new Error(`No view item for action '${action.id}'`);
-			},
-			orientation: ActionsOrientation.VERTICAL,
-			ariaLabel: localize('manage', "Manage"),
-			animated: false,
-			preventLoopNavigation: true
-		}));
-
-		this.globalActivityAction = this._register(new ActivityAction(this.createGlobalActivity()));
-		this._register(this.userDataProfileService.onDidChangeCurrentProfile(e => {
-			if (this.globalActivityAction) {
-				this.globalActivityAction.activity = this.createGlobalActivity();
+		this.keyboardNavigationDisposables.add(addDisposableListener(this.globalCompositeBar.element, EventType.KEY_DOWN, e => {
+			const kbEvent = new StandardKeyboardEvent(e);
+			if (kbEvent.equals(KeyCode.UpArrow) || kbEvent.equals(KeyCode.LeftArrow)) {
+				this.compositeBar?.focus(this.getVisiblePaneCompositeIds().length - 1);
 			}
 		}));
-
-		if (this.accountsVisibilityPreference) {
-			this.accountsActivityAction = this._register(new ActivityAction({
-				id: 'workbench.actions.accounts',
-				name: localize('accounts', "Accounts"),
-				classNames: ThemeIcon.asClassNameArray(ActivitybarPart.ACCOUNTS_ICON)
-			}));
-
-			this.globalActivityActionBar.push(this.accountsActivityAction, { index: ActivitybarPart.ACCOUNTS_ACTION_INDEX });
-		}
-
-		this.globalActivityActionBar.push(this.globalActivityAction);
-	}
-
-	private createGlobalActivity(): IActivity {
-		return {
-			id: 'workbench.actions.manage',
-			name: localize('manage', "Manage"),
-			classNames: ThemeIcon.asClassNameArray(this.userDataProfileService.currentProfile.icon ? ThemeIcon.fromId(this.userDataProfileService.currentProfile.icon) : DEFAULT_ICON),
-		};
-	}
-
-	private toggleAccountsActivity() {
-		if (!!this.accountsActivityAction === this.accountsVisibilityPreference) {
-			return;
-		}
-		if (this.globalActivityActionBar) {
-			if (this.accountsActivityAction) {
-				this.globalActivityActionBar.pull(ActivitybarPart.ACCOUNTS_ACTION_INDEX);
-				this.accountsActivityAction = undefined;
-			} else {
-				this.accountsActivityAction = this._register(new ActivityAction({
-					id: 'workbench.actions.accounts',
-					name: localize('accounts', "Accounts"),
-					classNames: ThemeIcon.asClassNameArray(Codicon.account)
-				}));
-				this.globalActivityActionBar.push(this.accountsActivityAction, { index: ActivitybarPart.ACCOUNTS_ACTION_INDEX });
-			}
-		}
-
-		this.updateGlobalActivity(ACCOUNTS_ACTIVITY_ID);
 	}
 
 	getPinnedPaneCompositeIds(): string[] {
@@ -448,18 +278,8 @@ export class ActivitybarPart extends Part {
 		if (this.menuBarContainer) {
 			availableHeight -= this.menuBarContainer.clientHeight;
 		}
-		if (this.globalActivityActionBar) {
-			availableHeight -= (this.globalActivityActionBar.viewItems.length * ActivitybarPart.ACTION_HEIGHT); // adjust height for global actions showing
-		}
+		availableHeight -= (this.globalCompositeBar.size() * ActivitybarPart.ACTION_HEIGHT); // adjust height for global actions showing
 		this.compositeBar.layout(width, availableHeight);
-	}
-
-	private get accountsVisibilityPreference(): boolean {
-		return this.storageService.getBoolean(AccountsActivityActionViewItem.ACCOUNTS_VISIBILITY_PREFERENCE_KEY, StorageScope.PROFILE, true);
-	}
-
-	private set accountsVisibilityPreference(value: boolean) {
-		this.storageService.store(AccountsActivityActionViewItem.ACCOUNTS_VISIBILITY_PREFERENCE_KEY, value, StorageScope.PROFILE, StorageTarget.USER);
 	}
 
 	toJSON(): object {
